Factored out most of the stuff from Cli
Cli really was an abomination. This feels much better.
