+++ src/main/java/minijava/ir/optimize/BaseOptimizer.java
@@ -48,7 +48,7 @@ public abstract class BaseOptimizer extends NodeVisitor.Default implements Optim
   }
 
   /** Just a helper method, Node.accept(NodeVisitor) flipped. */
-  protected void visit(Node node) {
+  protected final void visit(Node node) {
     node.accept(this);
   }
 }
