--- src/main/java/minijava/ir/optimize/ConsumingNodeVisitor.java
@@ -1,18 +1,0 @@
-package minijava.ir.optimize;
-
-import firm.nodes.*;
-import java.util.function.Consumer;
-
-/** A {@link NodeVisitor} that passes visited nodes to a {@link Consumer<Node>}. */
-public class ConsumingNodeVisitor extends NodeVisitor.Default {
-  private final Consumer<Node> consumer;
-
-  public ConsumingNodeVisitor(Consumer<Node> consumer) {
-    this.consumer = consumer;
-  }
-
-  @Override
-  public void defaultVisit(Node n) {
-    consumer.accept(n);
-  }
-}
